Scrapers/alphareadscraper/requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
supabase==1.0.3
httpx==0.23.0
gotrue==1.3.1
postgrest==0.10.6 